simpler loader race condition test
