+++ realm-server/tests/cards/a.js
@@ -1,0 +1,5 @@
+import { b } from "./b";
+
+export function a() {
+  return "a" + b();
+}
