+++ realm-server/tests/cards/b.js
@@ -1,0 +1,5 @@
+import { c } from "./c";
+
+export function b() {
+  return "b" + c();
+}
