+++ realm-server/tests/cards/c.js
@@ -1,0 +1,3 @@
+export function c() {
+  return "c";
+}
